clean descriptions from locus search terms
When the description terms of the locus search terms file are read, they will be automatically stripped of `,`, `:`, `)`, and `(` characters. This also happens with the description lines of the sequences from the fasta file, and allows search compatibility. This is a more robust check to the search terms, which should not depend on the above characters for their uniqueness.