add back compatibility for 3 column locus search terms files, fix isolate, strain, haplotype bug
New negative search terms feature requires a four-column search terms file. This breaks previous requirements for a three-column search terms file. Added a fix for this, when three columns are present the default value of 'N/A' automatically generated for the negative search term.

Also fixed a bug where if 'haplotype', 'isolate', or 'strain' is the only string present in the description portion (e.g., >ACC GENUS SPECIES isolate), an error was thrown. Generally these terms are present with a trailing code that identifies the sample further, but in some cases these words appear in isolation. This fix correctly identifies these cases and returns a null value for the voucher field.