increase BLAST parse speed
Change data structure created for parsing BLAST results. Instead of iterating through BLAST file per accession, create a dictionary with accession as key and value as list containing sublists of coord pairs. In testing, this produced a 4-8x speedup. The time-limiting factor for these scripts is now the BLASTn search itself, and not the coordinate merging step. This speedup is particularly noticeable for markers containing 10,000+ sequences.